WIP, snake not moving, food disappearing :'(
Ok, so Jamie did this with classes and I didn't know he was going to do
that. I tried to avoid them and refactor his code as the video was
playing, but I ended up with a broken game.

This will need a huge refactoring since I didn't follow Jamie's path.